audio_core/time_stretch: Silence truncation warnings in Process()
The SoundTouch API only accepts uint amount of samples.

+++ src/audio_core/time_stretch.cpp
@@ -62,8 +62,8 @@ std::size_t TimeStretcher::Process(const s16* in, std::size_t num_in, s16* out,
     LOG_DEBUG(Audio, "{:5}/{:5} ratio:{:0.6f} backlog:{:0.6f}", num_in, num_out, stretch_ratio,
               backlog_fullness);
 
-    sound_touch->putSamples(in, num_in);
-    return sound_touch->receiveSamples(out, num_out);
+    sound_touch->putSamples(in, static_cast<u32>(num_in));
+    return sound_touch->receiveSamples(out, static_cast<u32>(num_out));
 }
 
 void TimeStretcher::Clear() {
